Added event listeners to nav elements.
These event listeners are used to update which tab in the tidy5e sheet is clicked, which is then stored.

The other functions will update the HTML to correctly render the circles tab when necessary.

It seems to "blink" when an affinity is clicked, but this is still a big improvement.

Closes #2.